+++ embrace/lib/embrace.dart
@@ -7,6 +7,7 @@ import 'package:embrace/embrace_api.dart';
 import 'package:embrace/src/embrace_frame_detector.dart';
 import 'package:embrace/src/embrace_startup_tracker.dart';
 import 'package:embrace/src/otel/otel.dart';
+import 'package:embrace/src/pointer_input_tracker.dart';
 import 'package:embrace_platform_interface/embrace_platform_interface.dart';
 import 'package:embrace_platform_interface/last_run_end_state.dart';
 import 'package:flutter/widgets.dart';
@@ -549,6 +550,7 @@ Future<void> _start(
 ) async {
   WidgetsFlutterBinding.ensureInitialized();
   EmbraceStartupTracker.init();
+  EmbracePointerInputTracker.init();
 
   if (OTelFactory.otelFactory == null) {
     OTelAPI.initialize(