+++ embrace_dio/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 4.9.0
+
+* Version bump
+
+# 4.8.0
+
+* Version bump
+
 # 4.7.0
 
 * Version bump