+++ embrace_go_router/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 4.9.0
+
+* Added a screen-load span to `EmbraceGoRouterObserver`, timed from the last user touch to the first frame of the new route
+* Bug fixes
+
+# 4.8.0
+
+* Version bump
+
 # 4.7.0
 
 * Initial release of `embrace_go_router`. Provides `EmbraceGoRouterObserver` for automatic navigation tracking and time-to-interactive (TTI) span instrumentation in apps using go_router.